Improve the output of the build script
By adding the --fail, --silent and --show-error argument to the curl
command call the build script should be easier to debug when needed.
This patch also fixes shellcheck complains and adds some useful info
like which file is being downloaded and where it'll be extracted to.

This patch also contains a fix to display all output in the web
interface by making the RPC redirect stderr and stdout.